python/b32/pyproject.toml
[project]
name = "b32"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django-ninja>=1.4.3",
    "requests>=2.32.5",
]
